Cleanup message list
