+++ src/frontend/apps/dashboard/src/slices/product/pages/(logs)/provider-error/_layout.tsx
@@ -1,15 +1,16 @@
 import { renderWithLoader } from '@metorial/data-hooks';
 import { Paths } from '@metorial/frontend-config';
-import { ContentLayout, PageHeader } from '@metorial/layout';
+import { ContentLayout, ExtraHeaderLayout, PageHeader } from '@metorial/layout';
 import {
   useCurrentInstance,
   useCurrentOrganization,
   useCurrentProject,
   useSessionErrorGroup
 } from '@metorial/state';
-import { LinkTabs, RenderDate } from '@metorial/ui';
+import { Button, LinkTabs, RenderDate } from '@metorial/ui';
 import { ID } from '@metorial/ui-product';
-import { Outlet, useLocation, useParams } from 'react-router-dom';
+import { RiArrowLeftSLine } from '@remixicon/react';
+import { Link, Outlet, useLocation, useParams } from 'react-router-dom';
 import { AttributesLayout } from '../../../scenes/attributesLayout';
 
 export let ProviderErrorLayout = () => {
@@ -30,48 +31,68 @@ export let ProviderErrorLayout = () => {
   ] as const;
 
   return (
-    <ContentLayout>
-      <PageHeader
-        title={error.data?.message ?? ''}
-        pagination={[
-          {
-            label: 'Errors',
-            href: Paths.instance.providerErrors(organization.data, project.data, instance.data)
-          },
-          {
-            label: error.data?.code ?? 'Error',
-            href: Paths.instance.providerError(...providerPathParams)
-          }
-        ]}
-      />
+    <ExtraHeaderLayout
+      header={
+        <Link
+          to={Paths.instance.providerErrors(
+            organization.data,
+            project.data,
+            instance.data
+          )}
+        >
+          <Button size="2" variant="outline" iconLeft={<RiArrowLeftSLine />}>
+            Back to all errors
+          </Button>
+        </Link>
+      }
+    >
+      <ContentLayout>
+        <PageHeader
+          title={error.data?.message ?? ''}
+          pagination={[
+            {
+              label: 'Errors',
+              href: Paths.instance.providerErrors(
+                organization.data,
+                project.data,
+                instance.data
+              )
+            },
+            {
+              label: error.data?.code ?? 'Error',
+              href: Paths.instance.providerError(...providerPathParams)
+            }
+          ]}
+        />
 
-      <LinkTabs
-        current={pathname}
-        links={[
-          {
-            label: 'Occurrences',
-            to: Paths.instance.providerError(...providerPathParams)
-          }
-        ]}
-      />
-
-      {renderWithLoader({ error })(({ error }) => (
-        <AttributesLayout
-          variant="large"
-          items={[
-            { label: 'Error Group ID', value: <ID id={error.data.id} /> },
-            { label: 'First Seen', value: <RenderDate date={error.data.createdAt} /> },
+        <LinkTabs
+          current={pathname}
+          links={[
             {
               label: 'Occurrences',
-              value: error.data.occurrenceCount ?? '—'
-            },
-            { label: 'Code', value: error.data.code ?? '—' },
-            { label: 'Provider', value: error.data.providerId ?? '—' }
+              to: Paths.instance.providerError(...providerPathParams)
+            }
           ]}
-        >
-          <Outlet />
-        </AttributesLayout>
-      ))}
-    </ContentLayout>
+        />
+
+        {renderWithLoader({ error })(({ error }) => (
+          <AttributesLayout
+            variant="large"
+            items={[
+              { label: 'Error Group ID', value: <ID id={error.data.id} /> },
+              { label: 'First Seen', value: <RenderDate date={error.data.createdAt} /> },
+              {
+                label: 'Occurrences',
+                value: error.data.occurrenceCount ?? '—'
+              },
+              { label: 'Code', value: error.data.code ?? '—' },
+              { label: 'Provider', value: error.data.providerId ?? '—' }
+            ]}
+          >
+            <Outlet />
+          </AttributesLayout>
+        ))}
+      </ContentLayout>
+    </ExtraHeaderLayout>
   );
 };
