+++ src/frontend/apps/dashboard/src/slices/product/scenes/session-messages/components/argumentsView.tsx
@@ -1,0 +1,14 @@
+import { Text } from '@metorial/ui';
+import { JsonViewer } from '../../../../../components/jsonViewer';
+
+export let ArgumentsView = ({ args }: { args: Record<string, any> | null | undefined }) => {
+  if (!args || Object.keys(args).length === 0) {
+    return (
+      <Text size="1" color="gray700">
+        No arguments provided.
+      </Text>
+    );
+  }
+
+  return <JsonViewer value={args} />;
+};
