+++ src/frontend/apps/dashboard/src/slices/product/scenes/session-messages/components/contentBlocksView.tsx
@@ -1,0 +1,21 @@
+import { Text } from '@metorial/ui';
+import { BlockStack } from '../styles';
+import { ContentBlockView } from './contentBlockView';
+
+export let ContentBlocksView = ({ content }: { content: unknown }) => {
+  if (!Array.isArray(content) || content.length === 0) {
+    return (
+      <Text size="1" color="gray700">
+        No content returned.
+      </Text>
+    );
+  }
+
+  return (
+    <BlockStack>
+      {content.map((block, index) => (
+        <ContentBlockView key={index} content={block} index={index} />
+      ))}
+    </BlockStack>
+  );
+};
