+++ src/frontend/apps/dashboard/src/slices/product/scenes/session-messages/index.ts
@@ -1,0 +1,7 @@
+export { ExplorerCapabilitiesMessageGroup } from './components/explorerCapabilitiesMessageGroup';
+export { Message } from './components/message';
+export {
+  getMessageMethod,
+  isExplorerCapabilityMethod,
+  shouldRenderStandaloneMessage
+} from './utils';
