+++ src/frontend/apps/dashboard/src/slices/product/scenes/sessionTracing/components/connectionLogs.tsx
@@ -1,16 +1,18 @@
 import { DashboardInstanceSessionsGetOutput } from '@metorial/dashboard-sdk';
-import { Callout, CenteredSpinner, Entity, RenderDate } from '@metorial/ui';
+import { Callout, CenteredSpinner } from '@metorial/ui';
 import styled from 'styled-components';
-import { SessionConnectionStatusBadge } from '../../providerSessions/table';
 import { ItemList } from '../../session/components/itemList';
 import { useConnectionTimeline } from '../hooks/useConnectionTimeline';
 import { SessionConnection } from '../types';
+import { ConnectionTraceHeader } from './connectionTraceHeader';
 
 let DetailContent = styled.div`
   display: flex;
   flex-direction: column;
   gap: 16px;
   padding: 16px;
+  background: #fafafa;
+  min-height: 100%;
 `;
 
 let DetailTimeline = styled.div`
@@ -32,76 +34,27 @@ export let ConnectionLogs = ({
   session: DashboardInstanceSessionsGetOutput;
   connection: SessionConnection;
 }) => {
-  let { connectionName, hasTimelineActivity, isLoading, mcp, sessionEntry, timelineItems } =
-    useConnectionTimeline({
-      session,
-      connection
-    });
+  let {
+    connection: connectionDetails,
+    connectionProviders,
+    hasTimelineActivity,
+    isLoading,
+    mcp,
+    sessionEntry,
+    timelineItems
+  } = useConnectionTimeline({
+    session,
+    connection
+  });
 
   return (
     <DetailContent>
-      <Entity.Wrapper>
-        <Entity.Content>
-          <Entity.Field title="Connection" value={connectionName} />
-          <Entity.Field
-            title="Status"
-            value={
-              <SessionConnectionStatusBadge
-                connectionStatus={connection.connectionState}
-                hasErrors={connection.hasErrors}
-                hasWarnings={connection.hasWarnings}
-              />
-            }
-          />
-          <Entity.Field title="Connection ID" value={connection.id} />
-          <Entity.Field
-            title="Created At"
-            value={<RenderDate date={connection.createdAt} />}
-          />
-          {connection.lastActiveAt && (
-            <Entity.Field
-              title="Last Active"
-              value={<RenderDate date={connection.lastActiveAt} />}
-            />
-          )}
-        </Entity.Content>
-      </Entity.Wrapper>
-
-      {mcp &&
-        (mcp.client?.name ||
-          mcp.client?.version ||
-          mcp.server?.name ||
-          mcp.server?.version ||
-          mcp.connectionType) && (
-          <Entity.Wrapper>
-            <Entity.Content>
-              {(mcp.client?.name || mcp.client?.version) && (
-                <Entity.Field
-                  title="Client"
-                  value={[mcp.client?.name, mcp.client?.version].filter(Boolean).join('@')}
-                />
-              )}
-              {(mcp.server?.name || mcp.server?.version) && (
-                <Entity.Field
-                  title="Server"
-                  value={[mcp.server?.name, mcp.server?.version].filter(Boolean).join('@')}
-                />
-              )}
-              {mcp.connectionType && (
-                <Entity.Field
-                  title="Connected Via"
-                  value={
-                    {
-                      websocket: 'WebSocket',
-                      streamable_http: 'Streamable HTTP',
-                      sse: 'Server-Sent Events'
-                    }[mcp.connectionType] ?? mcp.connectionType
-                  }
-                />
-              )}
-            </Entity.Content>
-          </Entity.Wrapper>
-        )}
+      <ConnectionTraceHeader
+        connection={connectionDetails}
+        mcp={mcp}
+        providers={connectionProviders}
+        session={session}
+      />
 
       <DetailTimeline>
         {sessionEntry}
