+++ src/frontend/apps/dashboard/src/slices/product/scenes/sessionTracing/hooks/useSessionTracing.ts
@@ -455,6 +455,7 @@ export let useSessionTracing = (
     activeConnection,
     activeTabId,
     connectionCount: connectionItems.length,
+    connectionIdByExplorerTabId: explorerConnectionByTabId,
     explorerTabIdByConnectionId,
     explorerTabIds,
     groupedConnections,
